requirements.txt
# torch for training (latest version 2022.04.05)
pytorch == 1.11.0
torchvision == 0.12.0 